Homework 55

+++ src/main/java/backup/ServiceBackupRepository.java
@@ -9,7 +9,7 @@ public class ServiceBackupRepository<E extends Model> {
 
     private final List<E> repo = new ArrayList<>();
 
-    public List<E> createRepo(List<E> list, int courseID) {
+    public List<E> createRepo(List<E> list, long courseID) {
 //        return list.stream().filter(element -> element.getCourse().isPresent()
 //                && element.getCourse().get().getID() == courseID).toList();
         return null;
